phase16: add SYS_V2_DEBUG_WRITE_STR for atomic marker emission (WIP)
Added bounded atomic debug syscall to fix observability fragmentation:
- Syscall 1015: SYS_V2_DEBUG_WRITE_STR(str, length)
- Bounded copy (max 256 bytes)
- Atomic write (no interleaving)
- Proof/validation only (not general-purpose API)

Changes:
- Updated ABI: MAX_INDEX 14→15, NR 15→16
- Added kernel handler: sys_v2_debug_write_str()
- Added to enforcement matrix (Runtime_Bridge role)
- Updated ping payload to use atomic syscall
- Changed mask type: uint32_t→uint64_t (support syscall 15+)

STATUS: Implementation complete but test FAILS
- Payload calls syscall 1015
- But trace shows character fragmentation (still using PUTCHAR path?)
- 292 syscalls observed, marker not atomic
- Needs debug: handler reachability, enforcement validation

Task-5: INCOMPLETE - observability fix in progress

+++ kernel/sys/syscall_enforcement_matrix.c
@@ -21,7 +21,7 @@ static void enforcement_debug(const char *msg) {
  */
 int syscall_enforcement_validate(proc_execution_role_t role, uint64_t syscall_num) {
     const syscall_enforcement_entry_t *entry = NULL;
-    uint32_t syscall_mask;
+    uint64_t syscall_mask;
     
     /* Find enforcement entry for role */
     for (int i = 0; i < SYSCALL_ENFORCEMENT_MATRIX_SIZE; i++) {
@@ -90,7 +90,7 @@ const char* syscall_enforcement_get_role_name(proc_execution_role_t role) {
 /**
  * Get allowed syscall mask for role
  */
-uint32_t syscall_enforcement_get_allowed_mask(proc_execution_role_t role) {
+uint64_t syscall_enforcement_get_allowed_mask(proc_execution_role_t role) {
     for (int i = 0; i < SYSCALL_ENFORCEMENT_MATRIX_SIZE; i++) {
         if (SYSCALL_ENFORCEMENT_MATRIX[i].role == role) {
             return SYSCALL_ENFORCEMENT_MATRIX[i].allowed_syscalls_mask;
